[UI] HBTI 홈화면 하단 뷰 제거 및 리뷰 헤더 추가
