+++ HMOA_iOS/HMOA_iOS/Source/Presentation/Tab/Home/HBTI/HBTI/ViewController/HBTIViewController.swift
@@ -24,7 +24,7 @@ final class HBTIViewController: UIViewController, View {
     
     private let yourHBTIView = HBTIHomeTopView()
     
-    private let HBTIReviewHeaderView = HBTIHomeBottomView()
+    private let reivewHeader = HBTIHomeReviewHeaderView()
     
     // MARK: - Properties
     
@@ -85,7 +85,7 @@ final class HBTIViewController: UIViewController, View {
         view.addSubview(scrollView)
         
         [yourHBTIView, 
-         HBTIReviewHeaderView
+         reivewHeader
         ].forEach { scrollView.addSubview($0) }
         
     }
@@ -101,7 +101,7 @@ final class HBTIViewController: UIViewController, View {
             make.leading.trailing.equalToSuperview().inset(16)
         }
         
-        HBTIReviewHeaderView.snp.makeConstraints { make in
+        reivewHeader.snp.makeConstraints { make in
             make.top.equalTo(yourHBTIView.snp.bottom).offset(32)
             make.leading.trailing.equalToSuperview().inset(16)
             make.bottom.equalToSuperview().inset(20)
